+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.2.0-rc.2",
+  sdkVersion: "4.2.0-rc.3",
   genVersion: "2.788.15",
-  userAgent: "speakeasy-sdk/typescript 4.2.0-rc.2 2.788.15 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.2.0-rc.3 2.788.15 2.0 @orq-ai/node",
 } as const;