+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -576,7 +576,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KDSSBY6V7FMQS4VQWJ2TZHKX"),
+  _id: z.string().default("01KDT3BREG3CGPAN900T7C3F5F"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -619,7 +619,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KDSSBY6V7FMQS4VQWJ2TZHKX"),
+  id: z.string().default("01KDT3BREG3CGPAN900T7C3F5F"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,