+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4435,7 +4435,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KDSSBXQMCAN5W8C6FZDWD7DX"),
+  id: z.string().default("01KDT3BQZ3F3ZANVE05C5ZVD7T"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -4454,7 +4454,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KDSSBXQMCAN5W8C6FZDWD7DX"),
+  id: z.string().default("01KDT3BQZ3F3ZANVE05C5ZVD7T"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),