+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -307,7 +307,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kdssbxssxfx6j1dt951myk7b"),
+  _id: z.string().default("conv_01kdt3br1pm2355szjgxg5z4jc"),
   entityId: z.string(),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
@@ -341,7 +341,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kdssbxssxfx6j1dt951myk7b"),
+  id: z.string().default("conv_01kdt3br1pm2355szjgxg5z4jc"),
   entityId: z.string(),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),