+++ packages/orq-rc/RELEASES.md
@@ -5629,3 +5629,13 @@ Based on:
 - [typescript v4.2.0-rc.2] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.2 - packages/orq-rc
+
+## 2025-12-31 11:43:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.680.11 (2.788.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.3] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.3 - packages/orq-rc